Adding a script that allows my emacs config to be swapped out.
I orignially had a bunch of stuff baked into my bootstrap.org file,
but that was a little harder to maintain and it seems that there is
a new feature of emacs that breaks how that was working. Instead of
trying to get this working in the emacs config I opted to move my
symlinking outside of that context.