Fixing evil eol for up-list and forward-sexp optionally
Current approach gives the best of both worlds. The correct approach
for dealing with evil's emulate eol IMO is to address it at the root
with advice around forward-sexp and up-list. However, evil hacks around
this on a case-by-case basis. Which breaks my advice :(

I fixed this by only optionally applying my advice if the '+evil--should-correct-for-eol
variable is set to true. And then I advised my functions in +elisp.el to
set that to true. Ideally, I will comeback and figure out how to fix
evil to fix the problem at the root by undoing the hacks and putting the
advice at the bottom most level (forward-sexp & up-list), but for now this works for me