Adding pdf-view page-change conveniences
Specifically, adding `C-u`, `C-d` to change the page backward and
forward, respectively. Also, adding advice around going to the next page
directly. The default behavior preserves the vscroll of the current
page. That doesn't make much sense when reading to me, so move to the
top of the pdf image. I excluded that behavior from the backward page on
purpose -- I'm not sure what behavior I want in that case yet.