Adding support for cross-file src_block sharing
Not ready for 'prime-time', but I added the ability to temporarily load
the org files in a directory into the library of babel, so those blocks
can be used for noweb & src block calling. The imported blocks are
namespaced to the file. E.g.

If example/test.org has a source block called 'one', then you can
import it by referencing <<example/test:one>> or call it like normal