Updating 'global-packages to be faster
Previously it always called out to elpm. Since this is always hit during
startup performance was dragging. This makes it so we only call to elpm
automatically if there is no global-backages-load-paths.el file.

If updates are needed (more packages are going to be used) then you can
call #'global-packages/install-packages.