Adding ability to specify an org-export directory
Useful for me in working on homework assignments. I want
all of my assignments in org form in one directory and
I really don't care to see the .tex and .pdf in the same
directory. I only turn the assignment in once, so it's mostly
annoying. This gives me directory hygiene.