Enable JIT elisp eval in #+CALL and noweb references
I'm doing this with some black magic. Basic idea is,
when we resolve references stuff a temporary reference
in the library of babel with the elisp. Usage works like:

`#+CALL: (+ 2 2)`  or `<<(+ 2 2)>>`. You can still use header args (and
variables). For the gist the following will work:

`#+CALL: [:var a = 2](+ a <<(+ 3 3)>>)`