Rolled back to emacs28, but want my emojis.
This is a hacked backport of the emoji support
in emacs29. Functions are broken and some emojis
cannot be displayed / show up weird. Nothing I can do
about that for now.

I would like to get the emoji-insert working, but it's complaining
about recursion sadly. For now this is good enough.