Bringing in package that integrates with the system clipboard
Using the yank functionality forced a weird dependence on the
kill-ring, which varied depending on the instance of emacs running.
I use multiple emacs-daemons, so that hasn't been working for me. (ie
killing something in one emacs instance wasn't pulling the same value
to the top of another's kill ring) Soley using the system clipboard
resolves this. simpleclip accomplishes this.